Fixed signed vs unsigned comparison

+++ smarttools/NFmiLocalAreaMinMaxMask.cpp
@@ -857,8 +857,8 @@ static std::vector<LocalExtreme> LookForAdditionalLocalExtremes(
   ExtremeSearchCoreCounter searchCore;
 
   // Ei voi etsiä jos matriisi on liian pieni
-  if (subGridValues.NX() > 3 * searchCoreRadiusInGridPoints &&
-      subGridValues.NY() > 3 * searchCoreRadiusInGridPoints)
+  if (static_cast<long>(subGridValues.NX()) > 3 * searchCoreRadiusInGridPoints &&
+      static_cast<long>(subGridValues.NY()) > 3 * searchCoreRadiusInGridPoints)
   {
     // Käydään läpi kaikki ne pisteet mille voidaan tehdä laskut täydellä searchCore:lla
     for (size_t rowIndex = searchCoreRadiusInGridPoints;
